Switch to libevent for the TCP/HTTP server
This commit is a big one.

It switches the implementation of the server from the custom written tcp
server to a more performant solution, libevent.

The code has been inspired by the "http-server.c" example in the
"libevent/sample" folder, modified depending on what was needed.

Performance are acceptable in this revision, around 500 req/s with an
average of 8ms (16ms std deviation) per response, and the 99th percentile is
around 27ms.

Many things can be improved, but this is a good and solid starting point